fix: 무료설문 확인 후 스크리닝 삭제하도록 수정

+++ src/main/java/OneQ/OnSurvey/domain/survey/service/command/SurveyCommandService.java
@@ -325,7 +325,9 @@ public class SurveyCommandService implements SurveyCommand {
         surveyRepository.save(survey);
         surveyInfoRepository.save(info);
         surveyGlobalStatsService.addDueCount(info.getDueCount());
-        screeningRepository.deleteBySurveyId(surveyId);
+        if (survey.getIsFree()) {
+            screeningRepository.deleteBySurveyId(surveyId);
+        }
 
         applySurveyRuntimeCache(surveyId, userKey, dueCount, deadline);
